requirements.txt
RPi.GPIO>=0.7.0
spidev>=3.5
mfrc522>=0.0.7
smbus2>=0.4.0
RPLCD>=1.3.0
flask>=2.3.0
requests>=2.28.0
gpiozero>=1.6.0
